move filters panel logic to a component
